Gathering data for report - currently, results for the part 1 dataset analysis and the part 3 latency measurement have been gathered in textfiles in their respective folders.

+++ Lab6/part3/sine_approx_hfl.c
@@ -173,9 +173,9 @@ int main(void){
 			fl_result = 0.0;
 			fl_fact = 1.0;
 			for(int i = 0; i <= order; i++){
-				if(i!=0) fl_fact *= (2.0*i)*(2.0*i+1.0);
+				if(i!=0) fl_fact *= (float)(2.0*i)*(float)(2.0*i+1.0);
 				if(i % 2 == 0){
-					fl_result += (pow(x_fl, 2.0*i+1.0)/fl_fact);
+					fl_result += ((float)pow(x_fl, 2.0*i+1.0)/fl_fact);
 				}else{
 					fl_result -= (pow(x_fl, 2.0*i+1.0)/fl_fact);
 				}
